+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
+## v3.1.1 2018-05-22
+
+* Support `connect(port, host)` method arguments, again.
+
 ## v3.1.0 2018-05-19
 
-* Support `connect(port, host)` method arguments.
+* Support `connect(port, host)` method arguments (not really).
 
 ## v3.0.2 2018-03-12